feat: Update react-native-gesture-handler to version 2.16.2
This commit updates the react-native-gesture-handler dependency to version 2.16.2. This update includes bug fixes and performance improvements. It also adds new dependencies such as @egjs/hammerjs and hoist-non-react-statics. This update is necessary to ensure compatibility with the latest version of React Native and to take advantage of the new features and improvements provided by the updated library.