feat: Add support for background audio and camera permissions
This commit updates the app.json file to add support for background audio and camera permissions on iOS and Android respectively. This is necessary for the app to play sounds in the background and access the device's camera. These permissions enhance the functionality of the app and improve the user experience.