refactor: Update App component to use separate pages for home and wake up screens
This commit refactors the App component to use separate pages for the home screen and wake up screen. The home screen is displayed when the status is "none", and the wake up screen is displayed when the status is not "none". This change improves the organization of the code and enhances the user experience by providing distinct screens for different app states.