refactor: Add status property to AlarmStore
This commit adds a new `status` property to the `AlarmStore` interface and the `useAlarmStore` hook. The `status` property represents the current status of the alarm system and can have values such as "none", "active", or "snooze". This change enhances the functionality of the alarm system by providing a way to track and display the current status of alarms.